more a few method from Extensions to IAutoMapperBuilder
