+++ src/AutoMapper.Extensions.Microsoft.DependencyInjection/IAutoMapperBuilder.cs
@@ -1,4 +1,7 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
+using System;
+using System.Collections.Generic;
+using System.Reflection;
 using System.Text;
 
 namespace AutoMapper.Extensions.Microsoft.DependencyInjection
@@ -7,5 +10,9 @@ namespace AutoMapper.Extensions.Microsoft.DependencyInjection
     {
         IServiceCollection ServiceDescriptor { get; }
 
+        IAutoMapperBuilder AddAssemblies(IEnumerable<Assembly> assemblies);
+        IAutoMapperBuilder AddProfileTypes(IEnumerable<Type> profileTypes);
+
+        IAutoMapperBuilder AddRalatedTypes(IEnumerable<Type> ralatedTypes);
     }
 }
