+++ src/AutoMapper.Extensions.Microsoft.DependencyInjection/IPostAutoMapperConfiguration.cs
@@ -4,6 +4,9 @@ using System.Text;
 
 namespace AutoMapper.Extensions.Microsoft.DependencyInjection
 {
+    /// <summary>
+    /// represents an <see cref="Action{T1, T2}"/> that AutoMapper should care
+    /// </summary>
     public interface IPostAutoMapperConfiguration
     {
         void Configuration(IServiceProvider provider, IMapperConfigurationExpression configurationExpression);
