+++ src/AutoMapper.Extensions.Microsoft.DependencyInjection/Internal/AutoMapperConfigurationProvider.cs
@@ -40,13 +40,5 @@ namespace AutoMapper.Extensions.Microsoft.DependencyInjection.Internal
                 .Union(_typeDescriptions.SelectMany(x=>x.ProfileTypes));
         }
 
-        
-
-
-        //private static bool ImplementsGenericInterface(this Type type, Type interfaceType)
-        //    => type.IsGenericType(interfaceType) || type.GetTypeInfo().ImplementedInterfaces.Any(@interface => @interface.IsGenericType(interfaceType));
-
-        //private static bool IsGenericType(this Type type, Type genericType)
-        //    => type.GetTypeInfo().IsGenericType && type.GetGenericTypeDefinition() == genericType;
     }
 }
