+++ src/AutoMapper.Extensions.Microsoft.DependencyInjection/Internal/IAutoMapperTypeDescription.cs
@@ -4,6 +4,9 @@ using System.Text;
 
 namespace AutoMapper.Extensions.Microsoft.DependencyInjection.Internal
 {
+    /// <summary>
+    /// represents some <see cref="Profile"/> type that AutoMapper should care
+    /// </summary>
     public interface IAutoMapperTypeDescription
     {
         IEnumerable<Type> ProfileTypes { get; }
